Add Base. modules with common implementation of the elements
