+++ lib/AutoDemand/Elements/Source.ex
@@ -1,24 +1,43 @@
 defmodule AutoDemand.Elements.Source do
   use Membrane.Source
 
-  alias Membrane.Buffer
+  def_output_pad :output, caps: :any, demand_unit: :buffers
 
-  @message :crypto.strong_rand_bytes(1000)
-
-  def_output_pad :output, caps: :any
+  def_options initial_lower_bound: [
+                type: :integer,
+                spec: pos_integer,
+                description:
+                  "Initial lower bound for binsearching of the message generator frequency"
+              ],
+              initial_upper_bound: [
+                type: :integer,
+                spec: pos_integer,
+                description:
+                  "Initial upper bound for binsearching of the message generator frequency"
+              ]
 
   @impl true
-  def handle_init(_opts) do
-    {:ok, %{}}
+  def handle_init(opts) do
+    Base.Source.handle_init(opts)
   end
 
   @impl true
-  def handle_demand(:output, size, :buffers, _ctx, state) do
-    buffers =
-      1..size
-      |> Enum.map(fn _i -> %Buffer{payload: @message, dts: Membrane.Time.monotonic_time()} end)
+  def handle_prepared_to_playing(ctx, state) do
+    Base.Source.handle_prepared_to_playing(ctx, state)
+  end
 
-    actions = [buffer: {:output, buffers}]
-    {{:ok, actions}, state}
+  @impl true
+  def handle_tick(:next_buffer_timer, ctx, state = %{status: :playing}) do
+    Base.Source.handle_tick(:next_buffer_timer, ctx, state)
+  end
+
+  @impl true
+  def handle_other(msg, ctx, state) do
+    Base.Source.handle_other(msg, ctx, state)
+  end
+
+  @impl true
+  def handle_demand(:output, _size, :buffers, _ctx, state) do
+    {:ok, state}
   end
 end
